chore: update requirements.txt to replace pkg-resources with setuptools
The requirements.txt file was modified to replace 'pkg-resources' with 'setuptools'. This change is necessary because 'pkg-resources' is not a package that should be included in requirements.txt as it is a part of setuptools metadata. Replacing it with 'setuptools' ensures proper dependency management and resolves any potential issues during the installation process.